Put the bucket config in the environment variables
The clamav depends on several config files, such as /etc/clamav/clamd.conf,
/etc/clamav/freshclam.conf

So I still keep the config.json file, but instead of reading from the config.json, I read from environment variables and write to the config.json for the mirror bucket. This is to avoid making big changes to the clamav underlying logic.